removing sequencer globals
